Moving the volume methods from the `utils` to the `signpod` package.
Those methods are now only used in the `signepod` package, therefore,
there is no reason for those methods to be imported from a different
package.
